+++ moll/__init__.py
@@ -1,1 +1,13 @@
-from .core import *
+import sys
+
+import jax
+
+DEBUG = sys.gettrace() is not None
+TEST = "pytest" in sys.modules
+JIT = not DEBUG
+
+# JIT = False
+
+# TODO: not working with pytest-cov
+if not JIT:
+    jax.config.update("jax_disable_jit", True)
