+++ moll/picking/__init__.py
@@ -1,0 +1,3 @@
+from .online_picker import OnlineDiversityPicker
+
+__all__ = ["OnlineDiversityPicker"]
