+++ moll/picking/online_add_test.py
@@ -1,54 +1,13 @@
+import jax
 import jax.numpy as jnp
 import pytest
 
-from moll.core.distance import (
-    _add_points_to_bag,
-    add_point_to_bag,
-    euclidean,
-    finalize_updates,
-    matrix_cross_sum,
-    needless_point_idx,
-    tanimoto,
-)
-
-
-# Tanimoto tests
-@pytest.mark.parametrize(
-    "dtype",
-    [
-        bool,
-        jnp.bool_,
-        int,
-        jnp.int8,
-        jnp.int16,
-        jnp.uint8,
-        jnp.uint16,
-    ],
-)
-def test_tanimoto(dtype):
-    a = jnp.array([1, 1, 1, 1]).astype(dtype)
-    b = jnp.array([0, 0, 1, 1]).astype(dtype)
-    c = jnp.array([1, 1, 1, 0]).astype(dtype)
-    d = jnp.array([0, 0, 0, 0]).astype(dtype)
-
-    assert tanimoto(a, b) == 0.5
-    assert tanimoto(a, c) == 0.25
-    assert tanimoto(a, d) == 1.0
-    assert tanimoto(b, c) == 0.75
-    assert tanimoto(b, d) == 1.0
-    assert tanimoto(a, a) == 0.0
-    assert tanimoto(d, d) == 0.0
-
-
-# Euclidean tests
-def test_euclidean():
-    p1 = jnp.array([1, 2, 3])
-    p2 = jnp.array([4, 5, 6])
-    expected_result = jnp.sqrt(27)
-    assert euclidean(p1, p2) == expected_result
-
+from .online_add import _add_point, _finalize_updates, _needless_point_idx, add_points
+from ..metrics.metrics import euclidean
 
 # Test finalize_updates
+
+finalize_updates = jax.jit(_finalize_updates)
 
 
 @pytest.mark.parametrize(
@@ -82,6 +41,10 @@ def dummy_potential_fn(dist):
 
 # Test needless point search
 
+needless_point_idx = jax.jit(
+    _needless_point_idx, static_argnames=["dist_fn", "potential_fn"]
+)
+
 
 @pytest.mark.parametrize(
     "array, expected",
@@ -98,23 +61,15 @@ def test_find_needless_point(array, expected):
     assert idx == expected
 
 
-@pytest.mark.parametrize(
-    "matrix, i, j, expected",
-    [
-        ([[1, 0], [0, 2]], 0, 0, 1),
-        ([[1, 0], [0, 2]], 1, 1, 2),
-        ([[1, 2], [3, 4]], 0, 0, 6),
-        ([[1, 2], [3, 4]], 1, 1, 9),
-        ([[1, 3, 4], [3, 2, 5], [4, 5, 3]], 0, 0, 15),
+# Test add points
+
+add_point = jax.jit(
+    _add_point,
+    static_argnames=[
+        "dist_fn",
+        "k_neighbors",
     ],
 )
-def test_matrix_cross_sum(matrix, i, j, expected):
-    matrix = jnp.array(matrix)
-    result = matrix_cross_sum(matrix, i, j)
-    assert result == expected
-
-
-# Test add points
 
 
 @pytest.fixture
@@ -131,9 +86,9 @@ def X():
     )
 
 
-def test_add_point_to_bag(X):
+def test_add_point(X):
     x = jnp.array([4.3, 4.3])
-    X_updated, is_accepted, updated_idx = add_point_to_bag(
+    X_updated, is_accepted, updated_idx = add_point(
         x=x,
         X=X,
         dist_fn=euclidean,
@@ -161,11 +116,12 @@ def test_add_point_to_bag(X):
         ),
     ],
 )
-def test_add_points_to_bag(X, xs, acc_mask):
+def test_add_points(X, xs, acc_mask):
     xs = jnp.array(xs)
     acc_mask = jnp.array(acc_mask)
+    X_copy = X.copy()
 
-    _updated_idxs, X_updated, acceptance_mask = _add_points_to_bag(
+    _updated_idxs, X_updated, acceptance_mask = add_points(
         X=X,
         xs=xs,
         dist_fn=euclidean,
@@ -176,6 +132,6 @@ def test_add_points_to_bag(X, xs, acc_mask):
     )
 
     assert (acceptance_mask == acc_mask).all()
-    assert (X[:3] == X_updated[:3]).all()
+    assert (X_copy[:3] == X_updated[:3]).all()
     idx = jnp.argmax(acc_mask)
     assert (X_updated[4] == xs[idx]).all()
